requirements_test.txt
django==1.8.3

djangorestframework==3.1.3
djangorestframework-jwt==1.6.0

django-cors-headers==1.1.0
django-extensions==1.5.5
django-filter==0.10.0
django-permission==0.8.8
djangorestframework-composed-permissions==0.1

markdown==2.6.2
werkzeug==0.10.4
django-debug-toolbar==1.3.2
mock==1.3.0
